add feature for program register

+++ lib/models/common/program/program.dart
@@ -30,7 +30,7 @@ class Program {
 
 @JsonSerializable()
 class AverageRating {
-  int average;
+  double average;
   int count;
 
   AverageRating({
